Fix oooseqid query (#1)
When querying the pending_data table for changes that already exist for
our new row version, we should also check that the tablename matches,
not just the ctid. Two different tables can of course have rows with
identical ctids. This is difficult to trigger (without a large amount of
data) or predict, though, so I'm not sure how to write a test for it. I
did verify that it stopped all warnings (see the "This should never
happen!" comment) when trying to delete all raw release tags for a
particular user in a copy of the production database; without this
change the warning would trigger almost non-stop, and with this change
the deletion is completely silent.